adicionando efeitos de som
