+++ Animation Studies/Assets/Animations/Explode.anim
@@ -1615,6 +1615,13 @@ AnimationClip:
   m_HasGenericRootTransform: 1
   m_HasMotionFloatCurves: 0
   m_Events:
+  - time: 0
+    functionName: PlayInflateSound
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 1.9333333
     functionName: PlayExplodeParticle
     data:
