--- Animation Studies/Assets/Scripts/CallParticles.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 51dcfb18f82f4db4cb2d94674bca11bc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
