+++ Animation Studies/Assets/sfx.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fce275d0e3b8dce478fe469df1f220e0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
